+++ platforms/bukkit/common/src/main/java/com/dfsek/terra/bukkit/nms/Initializer.java
@@ -54,10 +54,6 @@ public interface Initializer {
     private static Initializer constructInitializer() {
         try {
             String packageVersion = NMS;
-            if (NMS.equals("v1_21_4")) {
-                packageVersion = "v1_21_3";
-            }
-
             Class<?> initializerClass = Class.forName(TERRA_PACKAGE + "." + packageVersion + ".NMSInitializer");
             try {
                 return (Initializer) initializerClass.getConstructor().newInstance();
