--- platforms/bukkit/nms/v1_21_3/src/main/java/com/dfsek/terra/bukkit/nms/v1_21_3/NMSBiomeInfo.java
@@ -1,10 +1,0 @@
-package com.dfsek.terra.bukkit.nms.v1_21_3;
-
-import net.minecraft.resources.ResourceKey;
-import net.minecraft.world.level.biome.Biome;
-
-import com.dfsek.terra.api.properties.Properties;
-
-
-public record NMSBiomeInfo(ResourceKey<Biome> biomeKey) implements Properties {
-}
