+++ platforms/bukkit/nms/v1_21_5/build.gradle.kts
