+++ platforms/bukkit/nms/v1_21_5/src/main/java/com/dfsek/terra/bukkit/nms/v1_21_5/NMSBiomeInfo.java
@@ -1,0 +1,10 @@
+package com.dfsek.terra.bukkit.nms.v1_21_5;
+
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.level.biome.Biome;
+
+import com.dfsek.terra.api.properties.Properties;
+
+
+public record NMSBiomeInfo(ResourceKey<Biome> biomeKey) implements Properties {
+}
