+++ platforms/bukkit/nms/v1_21_5/src/main/java/com/dfsek/terra/bukkit/nms/v1_21_5/config/SpawnCostConfig.java
@@ -1,0 +1,38 @@
+package com.dfsek.terra.bukkit.nms.v1_21_5.config;
+
+import com.dfsek.tectonic.api.config.template.annotations.Default;
+import com.dfsek.tectonic.api.config.template.annotations.Value;
+import com.dfsek.tectonic.api.config.template.object.ObjectTemplate;
+import net.minecraft.world.entity.EntityType;
+
+
+public class SpawnCostConfig implements ObjectTemplate<SpawnCostConfig> {
+    @Value("type")
+    @Default
+    private EntityType<?> type = null;
+    
+    @Value("mass")
+    @Default
+    private Double mass = null;
+    
+    @Value("gravity")
+    @Default
+    private Double gravity = null;
+    
+    public EntityType<?> getType() {
+        return type;
+    }
+    
+    public Double getMass() {
+        return mass;
+    }
+    
+    public Double getGravity() {
+        return gravity;
+    }
+    
+    @Override
+    public SpawnCostConfig get() {
+        return this;
+    }
+}
